Move the main light curve network block to a separate file
